# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EduPlatform — Nền tảng Giáo dục MVP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses/{id}/curriculum:
    get:
      operationId: getCourseCurriculum
      tags: [courses]
      summary: Get full curriculum tree (chapters > lessons) with progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Curriculum tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseCurriculum"
  /courses/{id}/chapters:
    get:
      operationId: listChapters
      tags: [courses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
    post:
      operationId: createChapter
      tags: [courses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChapterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /chapters/{chapterId}:
    patch:
      operationId: updateChapter
      tags: [courses]
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChapterBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
    delete:
      operationId: deleteChapter
      tags: [courses]
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /courses/{id}/chapters/reorder:
    post:
      operationId: reorderChapters
      tags: [courses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderBody"
      responses:
        "204":
          description: Reordered
  /chapters/{chapterId}/lessons:
    get:
      operationId: listLessons
      tags: [courses]
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [courses]
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [courses]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson with blocks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
    patch:
      operationId: updateLesson
      tags: [courses]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLessonBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [courses]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /chapters/{chapterId}/lessons/reorder:
    post:
      operationId: reorderLessons
      tags: [courses]
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderBody"
      responses:
        "204":
          description: Reordered
  /lessons/{lessonId}/blocks:
    get:
      operationId: listLessonBlocks
      tags: [courses]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LessonBlock"
    post:
      operationId: createLessonBlock
      tags: [courses]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBlockBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonBlock"
  /blocks/{blockId}:
    patch:
      operationId: updateLessonBlock
      tags: [courses]
      parameters:
        - name: blockId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLessonBlockBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonBlock"
    delete:
      operationId: deleteLessonBlock
      tags: [courses]
      parameters:
        - name: blockId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /lessons/{lessonId}/blocks/reorder:
    post:
      operationId: reorderLessonBlocks
      tags: [courses]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderBody"
      responses:
        "204":
          description: Reordered
  /lessons/{lessonId}/complete:
    post:
      operationId: markLessonComplete
      tags: [courses]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonProgressResult"

  # ─── Public Catalog & Enrollment ─────────────────────────────────────────
...
components:
  schemas:
...
    Chapter:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        orderIndex:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - courseId
        - title
        - orderIndex
        - createdAt
    CreateChapterBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      required:
        - title
    UpdateChapterBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        orderIndex:
          type: integer
    ReorderBody:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
      required:
        - ids
    Lesson:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        title:
          type: string
        slug:
          type: ["string", "null"]
        orderIndex:
          type: integer
        includeInPreview:
          type: boolean
        durationMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - chapterId
        - title
        - orderIndex
        - includeInPreview
        - createdAt
    CreateLessonBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        includeInPreview:
          type: boolean
        durationMinutes:
          type: integer
      required:
        - title
    UpdateLessonBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        includeInPreview:
          type: boolean
        durationMinutes:
          type: integer
        orderIndex:
          type: integer
    LessonBlock:
      type: object
      properties:
        id:
          type: integer
        lessonId:
          type: integer
        type:
          type: string
          description: "heading | text | youtube | upload | quiz | assignment"
        data:
          type: object
          additionalProperties: true
          description: "Type-specific block payload (free-form JSON)"
        orderIndex:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - lessonId
        - type
        - orderIndex
        - createdAt
...
    LessonDetail:
      allOf:
        - $ref: "#/components/schemas/Lesson"
        - type: object
          properties:
            blocks:
              type: array
              items:
                $ref: "#/components/schemas/LessonBlock"
            chapterTitle:
              type: ["string", "null"]
            courseId:
              type: integer
            completed:
              type: boolean
          required:
            - blocks
            - courseId
            - completed
    CourseCurriculum:
      type: object
      properties:
        courseId:
          type: integer
        totalLessons:
          type: integer
        completedLessons:
          type: integer
        progressPercent:
          type: integer
        chapters:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              title:
                type: string
              description:
                type: ["string", "null"]
              orderIndex:
                type: integer
              lessons:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    title:
                      type: string
                    slug:
                      type: ["string", "null"]
                    orderIndex:
                      type: integer
                    includeInPreview:
                      type: boolean
                    durationMinutes:
                      type: ["integer", "null"]
                    completed:
                      type: boolean
                  required:
                    - id
                    - title
                    - orderIndex
                    - includeInPreview
                    - completed
            required:
              - id
              - title
              - orderIndex
              - lessons
      required:
        - courseId
        - totalLessons
        - completedLessons
        - progressPercent
        - chapters
...